+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 - Function `getJointCountMatrix` to conos obect
 - New possibilities to customize the output of `saveConosForScanPy` 
 - Function `parseCellGroups` to parse properly cell groupings depending on user settings
+- Function `estimteWeightEntropyPerCell` to visualize alignment quality per cell
 
 ## [1.1.2] - 2019-07-16